interface tweaks

+++ www/DynamicData/PageTemplates/Details.aspx.cs
@@ -36,57 +36,54 @@ namespace DynamicData
                WeavverMaster.Width = "100%";
                ScriptManager.RegisterStartupScript(UpdatePanel1, typeof(string), "RunScripts", "run();", true);
 
-               FormView1.ModeChanged += new EventHandler(FormView1_ModeChanged);
-               FormView1.DataBound += new EventHandler(FormView1_DataBound);
-
                table = DynamicDataRouteHandler.GetRequestMetaTable(Context);
+               
+               if (LoggedInUser != null &&
+                    LoggedInUser.OrganizationId != new Guid(ConfigurationManager.AppSettings["default_organizationid"]))
+                    DetailsDataSource.WhereParameters.Add(new Parameter("OrganizationId", DbType.Guid, SelectedOrganization.Id.ToString()));
+               DetailsDataSource.EntityTypeFilter = table.EntityType.Name;
+
+               if (!IsPostBack)
                FormView1.SetMetaTable(table);
+
 
                IsPublic = true;
 
-               if (Request["IFrame"] == "true")
+               bool canListData = table.Provider.EntityType.CanListData().HasMatchingRole(GetUserRoles());
+               if (Request["IFrame"] != "true")
                {
                     DynamicHyperLink control = this.FindControlR<DynamicHyperLink>("BackToTheList");
                     if (control != null)
-                         control.Visible = false;
+                         control.Visible = canListData;
+               }
 
-                    if (SelectedOrganization != null &&
-                         LoggedInUser != null &&
-                         SelectedOrganization.OrganizationId == LoggedInUser.OrganizationId)
-                    {
-                         EditOptions.Visible = true;
-                    }
-                    else
-                    {
-                         EditOptions.Visible = false;
-                    }
+               AttachmentsList.Visible = false;
 
+               if (!IsPostBack)
+               {
                     if (Request["id"] == null)
                     {
                          SetTitle(null);
                          FormView1.ChangeMode(FormViewMode.Insert);
                          BeginEdit.Visible = false;
+                         CancelEdit.Visible = false;
                          DeleteItem.Visible = false;
                     }
                     else
                     {
-                         FormView1.ChangeMode(FormViewMode.ReadOnly);
+                         //FormView1.DataBind();
                     }
                }
-               DetailsDataSource.EntityTypeFilter = table.EntityType.Name;
-
-               if (LoggedInUser != null &&
-                    LoggedInUser.OrganizationId != new Guid(ConfigurationManager.AppSettings["default_organizationid"]))
-                    DetailsDataSource.WhereParameters.Add(new Parameter("OrganizationId", DbType.Guid, SelectedOrganization.Id.ToString()));
 
                //TODO: Add logic to let the model generate the title
                // -- for example to show: "Account item" vs "Bank of America Account"
                //Master.FormTitle = table.DisplayName + " entry";
 
+               // Add dynamic controls to the page footer
                if (FormView1.CurrentMode != FormViewMode.Insert)
                {
                     GenerateMenu(RowAction.Insert, table.Provider.EntityType);
-                    AttachmentsList.Visible = true;
+                    //AttachmentsList.Visible = true;
 
                     string projectionPath = "~/DynamicData/Projections/" + table.EntityType.ToString().Replace("Weavver.Data.", "") + ".ascx";
                     if (File.Exists(Server.MapPath(projectionPath)))
@@ -95,6 +92,15 @@ namespace DynamicData
                          Projections.Controls.Add(projection);
                     }
                }
+          }
+//-------------------------------------------------------------------------------------------
+          protected void Page_Load(object sender, EventArgs e)
+          {
+               ///DetailsDataSource.Include = table.ForeignKeyColumnsNames;
+          }
+//-------------------------------------------------------------------------------------------
+          protected void FormView1_ModeChanged(object sender, EventArgs e)
+          {
           }
 //-------------------------------------------------------------------------------------------
           void DetailsDataSource_Inserted(object sender, EntityDataSourceChangedEventArgs e)
@@ -108,47 +114,67 @@ namespace DynamicData
                string url = "Details.aspx?Id={0}&WindowId={1}&ParentId={2}&IFrame={3}";
                url = String.Format(url, auditData.Id, Request["WindowId"], Request["ParentId"], Request["IFrame"]);
 
-               ScriptManager.RegisterStartupScript(Page, this.GetType(), "refreshParent2", "<script type='text/javascript'>document.location.href = '" + url + "';</script>", false);
-               
-               //Response.Redirect(url);
+               if (Request["IFrame"] == "true")
+               { 
+                    ScriptManager.RegisterStartupScript(Page, this.GetType(), "refreshParent2", "<script type='text/javascript'>document.location.href = '" + url + "';</script>", false);
+               }
+               else
+               {
+                    Response.Redirect(url);
+               }
           }
 //-------------------------------------------------------------------------------------------
-          void FormView1_DataBound(object sender, EventArgs e)
+          protected void FormView1_DataBound(object sender, EventArgs e)
           {
                ICustomTypeDescriptor descriptor = FormView1.DataItem as ICustomTypeDescriptor;
                if (descriptor != null)
                {
-                    EntityObject owner = (EntityObject)descriptor.GetPropertyOwner(null);
-                //    object x = owner.
-               }
-          }
-//-------------------------------------------------------------------------------------------
-          void FormView1_ModeChanged(object sender, EventArgs e)
-          {
-               if (FormView1.CurrentMode == FormViewMode.Insert)
-               {
-                    SaveEdit.Visible = true;
-                    BeginEdit.Visible = false;
-                    DeleteItem.Visible = false;
-                    DetailsDataSource.EnableInsert = true;
-               }
+                    EntityObject entityObject = (EntityObject)descriptor.GetPropertyOwner(null);
 
-               if (FormView1.CurrentMode == FormViewMode.ReadOnly)
-               {
-                    BeginEdit.Visible = true;
-                    CancelEdit.Visible = false;
-               }
+                    string[] userRoles = GetUserRoles();
 
-               if (FormView1.CurrentMode == FormViewMode.Edit)
-               {
-                    SaveEdit.Visible = true;
-                    CancelEdit.Visible = true;
+                    DataAccess readPermissions = entityObject.ReadPermissions();
+                    FormView1.FindControlR<Literal>("Permissions").Text = "Accessible to: " + String.Join(", ", readPermissions.AllowedRoles); 
+
+                    DataAccess deletePermissions = entityObject.DeletePermissions();
+                    DeleteItem.Visible = (deletePermissions.HasMatchingRole(userRoles));
+                    DeleteItem.ToolTip = "Accessible to: " + String.Join(", ", deletePermissions.AllowedRoles);
+
+                    DataAccess updatePermissions = entityObject.UpdatePermissions();
+                    bool canUpdate = updatePermissions.HasAnyRole(userRoles);
+                    
+                    DataAccess insertPermissions = entityObject.InsertPermissions();
+                    bool canInsert = insertPermissions.HasAnyRole(userRoles);
+
+                    if (FormView1.CurrentMode == FormViewMode.ReadOnly)
+                    {
+                         BeginEdit.Visible = canUpdate;
+                         SaveEdit.Visible = false;
+                         CancelEdit.Visible = false;
+                    }
+
+                    if (FormView1.CurrentMode == FormViewMode.Insert)
+                    {
+                         SaveEdit.Visible = canInsert;
+                         BeginEdit.Visible = false;
+                         DeleteItem.Visible = false; // item doesn't exist so you can't delete it
+                    }
+
+                    if (canUpdate && FormView1.CurrentMode == FormViewMode.Edit)
+                    {
+                         if (updatePermissions.HasAnyRole(userRoles))
+                         {
+                              EditOptions.Visible = true;
+                              BeginEdit.Visible = false;
+                              SaveEdit.Visible = true;
+                              CancelEdit.Visible = true;
+                              UpdatePanel1.FindControlR<LinkButton>("BeginEdit").Visible = false;
+                              UpdatePanel1.FindControlR<LinkButton>("CancelEdit").Visible = true;
+                              DeleteItem.Visible = false;
+                         }
+                         BeginEdit.ToolTip = "Accessible to: " + String.Join(", ", updatePermissions.AllowedRoles);
+                    }
                }
-          }
-//-------------------------------------------------------------------------------------------
-          protected void Page_Load(object sender, EventArgs e)
-          {
-               ///DetailsDataSource.Include = table.ForeignKeyColumnsNames;
           }
 //-------------------------------------------------------------------------------------------
           protected string SetTitle(object dataItem)
@@ -170,34 +196,32 @@ namespace DynamicData
 //-------------------------------------------------------------------------------------------
           public void GenerateMenu(RowAction currentPage, Type type)
           {
+               string[] allowedRoles = GetUserRoles();
                MethodInfo[] methods = type.GetMethods();
                foreach (MethodInfo method in methods)
                {
                     foreach (object attrib in method.GetCustomAttributes(typeof(DynamicDataWebMethod), true))
                     {
                          DynamicDataWebMethod ddMethod = (DynamicDataWebMethod)attrib;
-                         foreach (string role in ddMethod.Roles)
+                         if (ddMethod.HasAnyRole(allowedRoles))
                          {
-                              if (User.IsInRole(role))
+                              LinkButton webMethod = new LinkButton();
+                              webMethod.ID = "DynamicWebMethod_" + method.Name;
+                              webMethod.Text = ((DynamicDataWebMethod)attrib).MethodName;
+                              webMethod.CommandName = method.Name;
+                              webMethod.Click += new EventHandler(DynamicWebMethod_Click);
+                              webMethod.CssClass = "attachmentLink";
+                              webMethod.ToolTip = "Accessible to: " + String.Join(", ", ddMethod.Roles);
+
+                              AvailableActions.Controls.Add(webMethod);
+
+                              if (ddMethod.RequiresPostback)
                               {
-                                   LinkButton webMethod = new LinkButton();
-                                   webMethod.ID = "DynamicWebMethod_" + method.Name;
-                                   webMethod.Text = ((DynamicDataWebMethod)attrib).MethodName;
-                                   webMethod.CommandName = method.Name;
-                                   webMethod.Click += new EventHandler(DynamicWebMethod_Click);
-                                   webMethod.CssClass = "attachmentLink";
-                                   webMethod.ToolTip = "Accessible to: " + String.Join(", ", ddMethod.Roles);
-
-                                   AvailableActions.Controls.Add(webMethod);
-
-                                   if (ddMethod.RequiresPostback)
-                                   {
-                                        PostBackTrigger trig1 = new PostBackTrigger();
-                                        trig1.ControlID = "DynamicWebMethod_" + method.Name;
-                                        UpdatePanel1.Triggers.Add(trig1);
-                                   }
-                                   break;
+                                   PostBackTrigger trig1 = new PostBackTrigger();
+                                   trig1.ControlID = "DynamicWebMethod_" + method.Name;
+                                   UpdatePanel1.Triggers.Add(trig1);
                               }
+                              break;
                          }
 
                          //if (attrib.GetType() == typeof(CSSAttribute))
@@ -207,7 +231,6 @@ namespace DynamicData
                     }
                }
 
-               
                ICustomTypeDescriptor descriptor = FormView1.DataItem as ICustomTypeDescriptor;
                if (descriptor != null)
                {
@@ -259,14 +282,8 @@ namespace DynamicData
                //Attachments.Controls.Add(litHR);
           }
 //-------------------------------------------------------------------------------------------
-          protected void FormView1_ItemCommand(object sender, FormViewCommandEventArgs e)
+          public void GetPermissions()
           {
-               if (e.CommandName == DataControlCommands.UpdateCommandName)
-               {
-               }
-               if (e.CommandName == DataControlCommands.CancelCommandName)
-               {
-               }
           }
 //-------------------------------------------------------------------------------------------
           protected void Save_Link(object sender, EventArgs e)
@@ -274,7 +291,8 @@ namespace DynamicData
                try
                {
                     FormView1.UpdateItem(true);
-                    ReadOnlyMode();
+                    FormView1.ChangeMode(FormViewMode.ReadOnly);
+                    FormView1.DataBind();
                     if (!String.IsNullOrEmpty(Request["ParentId"]))
                     {
                          ScriptManager.RegisterStartupScript(Page, this.GetType(), "refreshParent", "<script type='text/javascript'>parent.refreshParent('" + Request["ParentId"] + "');</script>", false);
@@ -289,38 +307,44 @@ namespace DynamicData
 //-------------------------------------------------------------------------------------------
           protected void Edit_Link(object sender, EventArgs e)
           {
-               //Response.Redirect("http://www.yahoo.com");
-
-               FormView1.ChangeMode(FormViewMode.Edit);
                FormView1.DataBind();
+               FormView1.ChangeMode(FormViewMode.Edit);
 
-               BeginEdit.Visible = false;
-               SaveEdit.Visible = true;
-               UpdatePanel1.FindControlR<LinkButton>("BeginEdit").Visible = false;
-               UpdatePanel1.FindControlR<LinkButton>("CancelEdit").Visible = true;
-               DeleteItem.Visible = false;
-          }
-//-------------------------------------------------------------------------------------------
-          protected void Delete_Link(object sender, EventArgs e)
-          {
-               FormView1.DeleteItem();
+               ICustomTypeDescriptor descriptor = FormView1.DataItem as ICustomTypeDescriptor;
+               if (descriptor != null)
+               {
+                    EntityObject owner = (EntityObject)descriptor.GetPropertyOwner(null);
+
+                    //Response.Redirect("http://www.yahoo.com");
+                    string[] userRoles = GetUserRoles();
+                    DataAccess updatePermissions = owner.UpdatePermissions();
+                    if (updatePermissions.HasAnyRole(userRoles))
+                    {
+                         FormView1.ChangeMode(FormViewMode.Edit);
+                         //FormView1.DataBind();
+                    }
+               }
           }
 //-------------------------------------------------------------------------------------------
           protected void Cancel_Link(object sender, EventArgs e)
           {
-               ReadOnlyMode();
+               FormView1.DataBind();
+               FormView1.ChangeMode(FormViewMode.ReadOnly);
           }
 //-------------------------------------------------------------------------------------------
-          private void ReadOnlyMode()
+          protected void Delete_Link(object sender, EventArgs e)
           {
-               FormView1.ChangeMode(FormViewMode.ReadOnly);
                FormView1.DataBind();
 
-               BeginEdit.Visible = true;
-               SaveEdit.Visible = false;
-               UpdatePanel1.FindControlR<LinkButton>("BeginEdit").Visible = true;
-               UpdatePanel1.FindControlR<LinkButton>("CancelEdit").Visible = false;
-               DeleteItem.Visible = true;
+               string[] userRoles = GetUserRoles();
+
+               ICustomTypeDescriptor descriptor = FormView1.DataItem as ICustomTypeDescriptor;
+               EntityObject entityObject = (EntityObject)descriptor.GetPropertyOwner(null);
+               DataAccess deletePermissions = entityObject.DeletePermissions();
+               if (deletePermissions.HasMatchingRole(userRoles))
+               {
+                    FormView1.DeleteItem();
+               }
           }
 //-------------------------------------------------------------------------------------------
           protected void FormView1_ItemDeleting(object sender, FormViewDeleteEventArgs e)
